Move notificationReducer to a separate file
